+++ src/test/builders/resident.builder.ts
@@ -1,6 +1,7 @@
 import { v4 as uuid } from 'uuid';
 import MonthYear from '_/extensions/date/month_year.extension';
 import { ContractResident } from '_/models/resident/contractResident';
+import { LinkedDocument } from '_/models/resident/document';
 import { RentInformation } from '_/models/resident/rent';
 import { Resident } from '_/models/resident/resident';
 import WaterMeterReading from '_/models/resident/water_meter_reading';
@@ -82,6 +83,11 @@ class ResidentBuilder {
 
   public withRentDeposit(rentDeposit: CurrencyInCents): ResidentBuilder {
     this.resident.rentDeposit = rentDeposit;
+    return this;
+  }
+
+  public addDocument(document: LinkedDocument): ResidentBuilder {
+    this.resident.documents.push(document);
     return this;
   }