+++ src/utils/validation/constraints.ts
@@ -33,6 +33,11 @@ export enum ValidationConstraint {
    * ValidationConstraint that checks if the tested value is defined
    */
   Defined,
+
+  /**
+   * ValidationConstraint that checks if the tested file value is defined
+   */
+  DefinedFile,
 }
 
 /**
@@ -54,6 +59,7 @@ export const CONSTRAINT_FUNCTIONS: {
   [ValidationConstraint.CurrencyWithZero]: currencyWithZeroConstraint,
   [ValidationConstraint.Defined]: definedConstraint,
   [ValidationConstraint.NoEmptyArray]: noEmptyArrayConstraint,
+  [ValidationConstraint.DefinedFile]: definedFileConstraint,
 };
 
 /**
@@ -61,6 +67,7 @@ export const CONSTRAINT_FUNCTIONS: {
  */
 export const ERROR_MESSAGES = {
   EMPTY: 'Darf nicht leer sein!',
+  NO_FILE: 'Datei ist erforderlich!',
   LTE_ZERO: 'Muss größer als 0 sein!',
   LT_ZERO: 'Muss mindestens 0 sein!',
   NO_INTEGER: 'Muss eine ganze Zahl sein!',
@@ -163,6 +170,16 @@ function monthConstraint(
 function definedConstraint(value: any): string | undefined {
   if (value === null || value === undefined) {
     return ERROR_MESSAGES.EMPTY;
+  }
+  return undefined;
+}
+
+/**
+ * Constraint function for the `DefinedFile` constraint
+ */
+function definedFileConstraint(value: any): string | undefined {
+  if (value === null || value === undefined) {
+    return ERROR_MESSAGES.NO_FILE;
   }
   return undefined;
 }