+++ src/views/ResidentView/ResidentInformation/DocumentInformation/DocumentInformation.tsx
@@ -1,12 +1,14 @@
 import { Button } from '@mui/material';
 import React, { useState } from 'react';
 import GenerateContractModal from './GenerateContractModal/GenerateContractModal';
+import UploadDocumentModal from './UploadDocumentModal/UploadDocumentModal';
 
 /**
  * Component that displays all documents linked to a resident
  */
 function DocumentInformation() {
   const [showContractGenerationModal, setShowContractGenerationModal] = useState(false);
+  const [showUploadDocumentModal, setShowUploadDocumentModal] = useState(false);
 
   return (
     <>
@@ -14,8 +16,15 @@ function DocumentInformation() {
         show={showContractGenerationModal}
         onClose={() => setShowContractGenerationModal(false)}
       />
+      <UploadDocumentModal
+        show={showUploadDocumentModal}
+        onCloseModal={() => setShowUploadDocumentModal(false)}
+      />
       <Button onClick={() => setShowContractGenerationModal(true)}>
         Vertrag generieren
+      </Button>
+      <Button onClick={() => setShowUploadDocumentModal(true)}>
+        Dokument hinzufügen
       </Button>
     </>
   );